requirements.txt
flask
python-dotenv
requests
huggingface_hub
